bug-1827506: move sym header extraction, rework tests
This moves extract_sym_file_header to a new tecken/libsym.py where it
can be used by the upload view and the tests.

This reworks the tests to use well-formed sym files in the tests/data/
directory.
