+++ tecken/tests/test_symbolstorage.py
@@ -37,8 +37,12 @@ def test_get_metadata(symbol_storage, lower_case_debug_id):
     debug_id = upload.debug_id
     if lower_case_debug_id:
         debug_id = debug_id.lower()
-    metadata = symbol_storage.get_metadata("xul.pdb", debug_id, "xul.sym")
+    metadata = symbol_storage.get_metadata(
+        "teckentest_js.pdb", debug_id, "teckentest_js.sym"
+    )
     assert metadata.download_url
+
+    # NOTE(willkg): this requests a symbol file that doesn't exist in storage
     assert not symbol_storage.get_metadata(
         "xxx.pdb", "44E4EC8C2F41492B9369D6B9A059577C2", "xxx.sym"
     )
