+++ tecken/tests/utils.py
@@ -6,10 +6,22 @@ from dataclasses import dataclass
 import gzip
 from hashlib import md5
 from io import BytesIO
+import os
 from typing import Optional
 
 from tecken.base.symbolstorage import SymbolStorage
 from tecken.libstorage import ObjectMetadata, StorageBackend
+from tecken.libsym import extract_sym_header_data
+
+
+def _load_sym_file(path: str):
+    path = os.path.join(os.path.dirname(__file__), "data", path)
+    with open(path, "rb") as fp:
+        body = fp.read()
+
+    data = extract_sym_header_data(path)
+    data["body"] = body
+    return data
 
 
 @dataclass
@@ -27,37 +39,36 @@ class Upload:
         return SymbolStorage.make_key(self.debug_file, self.debug_id, self.sym_file)
 
     @classmethod
-    def uncompressed(
-        cls, debug_file: str, debug_id: str, sym_file: str, body: bytes
-    ) -> "Upload":
-        metadata = ObjectMetadata(content_length=len(body))
+    def uncompressed(cls, sym_file: str) -> "Upload":
+        data = _load_sym_file(sym_file)
+
+        metadata = ObjectMetadata(content_length=len(data["body"]))
         return cls(
-            debug_file=debug_file,
-            debug_id=debug_id,
+            debug_file=data.get("debug_filename", ""),
+            debug_id=data.get("debug_id", ""),
             sym_file=sym_file,
-            body=body,
+            body=data["body"],
             metadata=metadata,
         )
 
     @classmethod
-    def compressed(
-        cls, debug_file: str, debug_id: str, sym_file: str, body: bytes
-    ) -> "Upload":
-        compressed_body = gzip.compress(body)
+    def compressed(cls, sym_file: str) -> "Upload":
+        data = _load_sym_file(sym_file)
+        compressed_body = gzip.compress(data["body"])
         metadata = ObjectMetadata(
             content_type="text/plain",
             content_length=len(compressed_body),
             content_encoding="gzip",
-            original_content_length=len(body),
-            original_md5_sum=md5(body).hexdigest(),
+            original_content_length=len(data["body"]),
+            original_md5_sum=md5(data["body"]).hexdigest(),
         )
         return cls(
-            debug_file=debug_file,
-            debug_id=debug_id,
+            debug_file=data.get("debug_filename", ""),
+            debug_id=data.get("debug_id", ""),
             sym_file=sym_file,
             body=compressed_body,
             metadata=metadata,
-            original_body=body,
+            original_body=data["body"],
         )
 
     def upload_to_backend(self, backend: StorageBackend):
@@ -70,19 +81,10 @@ class Upload:
 
 
 UPLOADS = {
-    "uncompressed": Upload.uncompressed(
-        debug_file="a", debug_id="b", sym_file="c", body=b"test"
-    ),
-    "compressed": Upload.compressed(
-        debug_file="xul.pdb",
-        debug_id="44E4EC8C2F41492B9369D6B9A059577C2",
-        sym_file="xul.sym",
-        body=b"symbols file contents",
-    ),
-    "special_chars": Upload.compressed(
-        debug_file="libc++abi.dylib",
-        debug_id="43940F08B65E38888CD3C52398EB1CA10",
-        sym_file="libc++abi.dylib.sym",
-        body=b"symbols file contents",
-    ),
+    # Linux executable
+    "uncompressed": Upload.uncompressed("teckentest_xpcshell.sym"),
+    # Windows pdb file
+    "compressed": Upload.compressed("teckentest_js.sym"),
+    # macOS dylib file
+    "special_chars": Upload.compressed("teckentest_libc++abi.dylib.sym"),
 }
